+++ src/strict-rules.ts
@@ -10,7 +10,9 @@ export const strictRules: DummyRuleMap = {
   "2h2d/require-promise-rejection-parameter": "error",
   "array-callback-return": "error",
   eqeqeq: ["error", "always", { null: "ignore" }],
+  "no-extend-native": "error",
   "no-new-func": "error",
+  "no-var": "error",
   "preserve-caught-error": ["error", { requireCatchParameter: true }],
   "typescript/ban-ts-comment": [
     "error",