+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Prohibit function-scoped `var` declarations through native `no-var`.
+- Prevent additions to built-in prototypes through native `no-extend-native`.
+
 ## [0.1.0-alpha.12] - 2026-08-24
 
 ### Fixed